add actions

+++ src/main/java/com/example/ReviewsInTheStore/services/impl/EmployeeServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.ReviewsInTheStore.services.impl;
 
 import com.example.ReviewsInTheStore.models.Assignment;
 import com.example.ReviewsInTheStore.models.Employee;
+import com.example.ReviewsInTheStore.models.Role;
 import com.example.ReviewsInTheStore.repositories.EmployeeRepository;
 import com.example.ReviewsInTheStore.services.EmployeeService;
 import com.example.ReviewsInTheStore.services.dtos.AssignmentView;
@@ -28,12 +29,6 @@ public class EmployeeServiceImpl implements EmployeeService {
         this.modelMapper = modelMapper;
     }
 
-//    @Override
-//    public EmployeeDTO createEmployee(EmployeeDTO employeeDto) {
-//        Employee employee = modelMapper.map(employeeDto, Employee.class);
-//        Employee savedEmployee = employeeRepository.save(employee);
-//        return modelMapper.map(savedEmployee, EmployeeDTO.class);
-//    }
 
     @Override
     public EmployeeView createEmployee(EmployeeView employeeView) {
@@ -43,8 +38,13 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override
-    public void deleteEmployee(UUID employeeId) {
+    public String deleteEmployee(UUID employeeId) {
+        Optional<Employee> employeeOpt = employeeRepository.findById(employeeId);
+        if (employeeOpt.isEmpty()){
+            return null;
+        }
         employeeRepository.deleteById(employeeId);
+        return "Employee deleted";
     }
 
     @Override
@@ -54,13 +54,7 @@ public class EmployeeServiceImpl implements EmployeeService {
         List<EmployeeView> employeeViews = new ArrayList<>();
         for (Employee employee : employees){
             EmployeeView employeeView = modelMapper.map(employee, EmployeeView.class);
-            List<AssignmentView> assignmentViews = new ArrayList<>();
-            List<Assignment> assignments = employee.getAssignments();
-            for (Assignment assignment: assignments){
-                assignmentViews.add(new AssignmentView(assignment.getId(),
-                        assignment.getFeedback().getId(), assignment.getAssignedTo().getId()));
-            }
-            employeeView.setAssignmentViewList(assignmentViews);
+            employeeView.setAssignmentViewList(getEmployeeAssignments(employee));
             employeeViews.add(employeeView);
         }
         return employeeViews;
@@ -74,13 +68,36 @@ public class EmployeeServiceImpl implements EmployeeService {
         }
         Employee employee = employeeOpt.get();
         EmployeeView employeeView = modelMapper.map(employee, EmployeeView.class);
+        employeeView.setAssignmentViewList(getEmployeeAssignments(employee));
+        return employeeView;
+    }
+
+
+    public EmployeeView changeRole(UUID id, String newRole) {
+        Optional<Employee> optionalEmployee = employeeRepository.findById(id);
+        if (optionalEmployee.isEmpty()) {
+            return null;
+        }
+        Employee employee = optionalEmployee.get();
+        try {
+                employee.setRole(Role.valueOf(newRole));
+            }
+        catch (IllegalArgumentException e) {
+            return null;
+        }
+
+        EmployeeView employeeView = modelMapper.map(employeeRepository.saveAndFlush(employee), EmployeeView.class);
+        employeeView.setAssignmentViewList(getEmployeeAssignments(employee));
+        return employeeView;
+    }
+
+    public List<AssignmentView> getEmployeeAssignments(Employee employee){
         List<AssignmentView> assignmentViews = new ArrayList<>();
         List<Assignment> assignments = employee.getAssignments();
         for (Assignment assignment: assignments){
             assignmentViews.add(new AssignmentView(assignment.getId(),
                     assignment.getFeedback().getId(), assignment.getAssignedTo().getId()));
         }
-        employeeView.setAssignmentViewList(assignmentViews);
-        return employeeView;
+        return assignmentViews;
     }
 }
